build: move to react-sbb-polarion 2.0.0
The topics app kept its own port of the breadcrumb bridge loader,
BreadcrumbTopic.tsx, for two reasons its doc comment named: the shared
component had no `parent` prop, so a sub-topic could not render
"Diff Tool > Collections", and it returned early when the loader script
was already in the shell head, so the first topic's label stuck for the
session. 2.0.0 fixes both, so the port goes and TopicsApp renders
BreadcrumbInjector - same four props, no adapter.

That also ends the runtime dependency on generic's webapp for the
bridge: the library ships it as a classic script, and the Vite build
copies it next to the built entries, where BreadcrumbInjector resolves
it from. configureGenericModules is removed in 2.0.0 and both entries
drop the call; genericModules.ts had no other reader.

The <link>s to generic's github-markdown-light.css and configurations.css
go too - 2.0.0 bundles both into style.css. The /ui/generic dev proxy
entry and the web.xml exemption stay: the three viewer pages and
SearchableSelect.jsx still load control CSS from there.

BreadcrumbTopic.test.tsx covered the port; the behaviour is the shared
component's now, so what remains repo-side - that a topic asks for a
breadcrumb, with its own title under its parent - moves into
TopicsApp.test.tsx.
