+++ ui/index.html
@@ -11,9 +11,6 @@
          NOT linked here - react-sbb-polarion bundles it into its own stylesheet, which main.tsx
          imports, with the icons inlined. -->
     <link rel="stylesheet" href="/polarion/wiki/skins/sidecar/presentation.css" />
-    <link rel="stylesheet" href="/polarion/diff-tool-app/ui/generic/css/github-markdown-light.css" />
-    <!-- Styles the named-configuration selector rendered by react-sbb-polarion's ConfigurationsPane. -->
-    <link rel="stylesheet" href="/polarion/diff-tool-app/ui/generic/css/configurations.css" />
   </head>
   <body class="sbb-ui">
     <div id="root"></div>
