+++ ui/vite.config.js
@@ -1,3 +1,5 @@
+import { copyFileSync } from 'node:fs';
+import { createRequire } from 'node:module';
 import { fileURLToPath, URL } from 'node:url';
 import react from '@vitejs/plugin-react';
 import { defineConfig, loadEnv } from 'vite';
@@ -34,8 +36,25 @@ export default defineConfig(({ command, mode }) => {
   const env = loadEnv(mode, process.cwd(), '');
   const polarionUrl = env.VITE_BASE_URL || 'http://localhost';
 
+// react-sbb-polarion's BreadcrumbInjector loads breadcrumb-bridge.js from next to the running page. It
+// runs in the Polarion shell window rather than in this app's frame, so it stays a classic script and
+// cannot be bundled - it is copied next to the built entries instead. See "Shell scripts" in the
+// library's README.
+function copyRspShellScripts() {
+  return {
+    name: 'copy-rsp-shell-scripts',
+    writeBundle(options) {
+      const require = createRequire(import.meta.url);
+      copyFileSync(
+        require.resolve('@sbb-polarion/react-sbb-polarion/breadcrumb-bridge.js'),
+        `${options.dir}/breadcrumb-bridge.js`,
+      );
+    },
+  };
+}
+
   const shared = {
-    plugins: [react()],
+    plugins: [react(), copyRspShellScripts()],
     resolve: {
       alias: { '@': resolvePath('./src') },
       // The app and (from stage 4 on) the linked react-sbb-polarion package must resolve to a single
@@ -66,11 +85,10 @@ export default defineConfig(({ command, mode }) => {
   };
 
   if (command === 'serve') {
-    // Everything the pages load from Polarion itself: REST, the generic UI toolkit CSS/JS, the wiki
+    // Everything the pages load from Polarion itself: REST, the wiki
     // skin stylesheet, and the icon/font assets referenced from CSS. Replaces the
     // NEXT_PUBLIC_BASE_URL prefixing useRemote used to do, so requests stay same-origin with no CORS.
     const polarionProxy = {
-      '/polarion/diff-tool-app/ui/generic': { target: polarionUrl, changeOrigin: true },
       '/polarion/diff-tool/rest': { target: polarionUrl, changeOrigin: true },
       '/polarion/diff-tool/ui': { target: polarionUrl, changeOrigin: true },
       '/polarion/wiki': { target: polarionUrl, changeOrigin: true },
